+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/CompleteConfigurationTest.php
@@ -18,7 +18,6 @@ use Symfony\Component\DependencyInjection\Argument\IteratorArgument;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\Security\Core\Authorization\AccessDecisionManager;
-use Symfony\Component\Security\Core\Encoder\Argon2iPasswordEncoder;
 use Symfony\Component\Security\Core\Encoder\SodiumPasswordEncoder;
 
 abstract class CompleteConfigurationTest extends TestCase
@@ -373,116 +372,6 @@ abstract class CompleteConfigurationTest extends TestCase
             'JMS\FooBundle\Entity\User7' => [
                 'class' => 'Symfony\Component\Security\Core\Encoder\SodiumPasswordEncoder',
                 'arguments' => [8, 128 * 1024 * 1024],
-            ],
-        ]], $container->getDefinition('security.encoder_factory.generic')->getArguments());
-    }
-
-    /**
-     * @group legacy
-     *
-     * @expectedDeprecation Configuring an encoder with "argon2i" as algorithm is deprecated since Symfony 4.3, use "auto" instead.
-     */
-    public function testEncodersWithArgon2i()
-    {
-        if (!Argon2iPasswordEncoder::isSupported()) {
-            $this->markTestSkipped('Argon2i algorithm is not supported.');
-        }
-
-        $container = $this->getContainer('argon2i_encoder');
-
-        $this->assertEquals([[
-            'JMS\FooBundle\Entity\User1' => [
-                'class' => 'Symfony\Component\Security\Core\Encoder\PlaintextPasswordEncoder',
-                'arguments' => [false],
-            ],
-            'JMS\FooBundle\Entity\User2' => [
-                'algorithm' => 'sha1',
-                'encode_as_base64' => false,
-                'iterations' => 5,
-                'hash_algorithm' => 'sha512',
-                'key_length' => 40,
-                'ignore_case' => false,
-                'cost' => null,
-                'memory_cost' => null,
-                'time_cost' => null,
-                'threads' => null,
-            ],
-            'JMS\FooBundle\Entity\User3' => [
-                'algorithm' => 'md5',
-                'hash_algorithm' => 'sha512',
-                'key_length' => 40,
-                'ignore_case' => false,
-                'encode_as_base64' => true,
-                'iterations' => 5000,
-                'cost' => null,
-                'memory_cost' => null,
-                'time_cost' => null,
-                'threads' => null,
-            ],
-            'JMS\FooBundle\Entity\User4' => new Reference('security.encoder.foo'),
-            'JMS\FooBundle\Entity\User5' => [
-                'class' => 'Symfony\Component\Security\Core\Encoder\Pbkdf2PasswordEncoder',
-                'arguments' => ['sha1', false, 5, 30],
-            ],
-            'JMS\FooBundle\Entity\User6' => [
-                'class' => 'Symfony\Component\Security\Core\Encoder\NativePasswordEncoder',
-                'arguments' => [8, 102400, 15],
-            ],
-            'JMS\FooBundle\Entity\User7' => [
-                'class' => 'Symfony\Component\Security\Core\Encoder\Argon2iPasswordEncoder',
-                'arguments' => [256, 1, 2],
-            ],
-        ]], $container->getDefinition('security.encoder_factory.generic')->getArguments());
-    }
-
-    /**
-     * @group legacy
-     */
-    public function testEncodersWithBCrypt()
-    {
-        $container = $this->getContainer('bcrypt_encoder');
-
-        $this->assertEquals([[
-            'JMS\FooBundle\Entity\User1' => [
-                'class' => 'Symfony\Component\Security\Core\Encoder\PlaintextPasswordEncoder',
-                'arguments' => [false],
-            ],
-            'JMS\FooBundle\Entity\User2' => [
-                'algorithm' => 'sha1',
-                'encode_as_base64' => false,
-                'iterations' => 5,
-                'hash_algorithm' => 'sha512',
-                'key_length' => 40,
-                'ignore_case' => false,
-                'cost' => null,
-                'memory_cost' => null,
-                'time_cost' => null,
-                'threads' => null,
-            ],
-            'JMS\FooBundle\Entity\User3' => [
-                'algorithm' => 'md5',
-                'hash_algorithm' => 'sha512',
-                'key_length' => 40,
-                'ignore_case' => false,
-                'encode_as_base64' => true,
-                'iterations' => 5000,
-                'cost' => null,
-                'memory_cost' => null,
-                'time_cost' => null,
-                'threads' => null,
-            ],
-            'JMS\FooBundle\Entity\User4' => new Reference('security.encoder.foo'),
-            'JMS\FooBundle\Entity\User5' => [
-                'class' => 'Symfony\Component\Security\Core\Encoder\Pbkdf2PasswordEncoder',
-                'arguments' => ['sha1', false, 5, 30],
-            ],
-            'JMS\FooBundle\Entity\User6' => [
-                'class' => 'Symfony\Component\Security\Core\Encoder\NativePasswordEncoder',
-                'arguments' => [8, 102400, 15],
-            ],
-            'JMS\FooBundle\Entity\User7' => [
-                'class' => 'Symfony\Component\Security\Core\Encoder\BCryptPasswordEncoder',
-                'arguments' => [15],
             ],
         ]], $container->getDefinition('security.encoder_factory.generic')->getArguments());
     }